+++ src/main/java/com/ships/repositories/ShippingCompanyRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ShippingCompanyRepository extends CrudRepository<ShippingCompany, Long>  {
+public interface ShippingCompanyRepository extends CrudRepository<ShippingCompany, Integer>  {
 	// Basic queries provided
 	// save(entity): save an Iterable of entities. Here, we can pass multiple objects to save them in a batch
 	// findAll(): get an Iterable of all available entities in database
